refactor: extract recommend page logic to custom hooks
